fix: add some dabian refactor to ensure the coverage can reach 95%
